LRQA-32801 Move to elements package
