+++ modules/test/poshi-runner/poshi-runner/src/main/java/com/liferay/poshi/runner/elements/BasePoshiElement.java
@@ -12,18 +12,18 @@
  * details.
  */
 
-package com.liferay.poshi.runner;
+package com.liferay.poshi.runner.elements;
 
-import static com.liferay.poshi.runner.ReadableSyntaxKeys.AND;
-import static com.liferay.poshi.runner.ReadableSyntaxKeys.BACKGROUND;
-import static com.liferay.poshi.runner.ReadableSyntaxKeys.FEATURE;
-import static com.liferay.poshi.runner.ReadableSyntaxKeys.GIVEN;
-import static com.liferay.poshi.runner.ReadableSyntaxKeys.SCENARIO;
-import static com.liferay.poshi.runner.ReadableSyntaxKeys.SETUP;
-import static com.liferay.poshi.runner.ReadableSyntaxKeys.TEARDOWN;
-import static com.liferay.poshi.runner.ReadableSyntaxKeys.THEN;
-import static com.liferay.poshi.runner.ReadableSyntaxKeys.THESE_PROPERTIES;
-import static com.liferay.poshi.runner.ReadableSyntaxKeys.WHEN;
+import static com.liferay.poshi.runner.elements.ReadableSyntaxKeys.AND;
+import static com.liferay.poshi.runner.elements.ReadableSyntaxKeys.BACKGROUND;
+import static com.liferay.poshi.runner.elements.ReadableSyntaxKeys.FEATURE;
+import static com.liferay.poshi.runner.elements.ReadableSyntaxKeys.GIVEN;
+import static com.liferay.poshi.runner.elements.ReadableSyntaxKeys.SCENARIO;
+import static com.liferay.poshi.runner.elements.ReadableSyntaxKeys.SETUP;
+import static com.liferay.poshi.runner.elements.ReadableSyntaxKeys.TEARDOWN;
+import static com.liferay.poshi.runner.elements.ReadableSyntaxKeys.THEN;
+import static com.liferay.poshi.runner.elements.ReadableSyntaxKeys.THESE_PROPERTIES;
+import static com.liferay.poshi.runner.elements.ReadableSyntaxKeys.WHEN;
 import static com.liferay.poshi.runner.util.StringPool.COLON;
 import static com.liferay.poshi.runner.util.StringPool.PIPE;
 import static com.liferay.poshi.runner.util.StringPool.TAB;
